Feat: udistrital/sga_documentacion#25 cambio de modelos
se cambian los modelos por helpers o utils segun el caso
