+++ controllers/generar_recibo.go
@@ -6,8 +6,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"strings"
-	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
@@ -15,7 +13,6 @@ import (
 	"github.com/phpdave11/gofpdf/contrib/barcode"
 	"github.com/udistrital/sga_mid_inscripcion/utils"
 	"github.com/udistrital/utils_oas/request"
-	"github.com/udistrital/sga_mid_inscripcion/utils"
 	"strings"
 	"time"
 )
