+++ controllers/solicitud_produccion.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
-	"github.com/udistrital/sga_mid_inscripcion/models"
+	"github.com/udistrital/sga_mid_inscripcion/helpers"
 )
 
 // SolicitudProduccionController ...
@@ -42,11 +42,11 @@ func (c *SolicitudProduccionController) PostAlertSolicitudProduccion() {
 	fmt.Println("Id Tercero: ", idTipoProduccionSrt)
 
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &SolicitudProduccion); err == nil {
-		if SolicitudProduccionAlert, errAlert := models.CheckCriteriaData(SolicitudProduccion, idTipoProduccion, idTercero); errAlert == nil {
-			if SolicitudProduccionPut, errCoincidence := models.CheckCoincidenceProduction(SolicitudProduccionAlert, idTipoProduccion, idTercero); errCoincidence == nil {
+		if SolicitudProduccionAlert, errAlert := helpers.CheckCriteriaData(SolicitudProduccion, idTipoProduccion, idTercero); errAlert == nil {
+			if SolicitudProduccionPut, errCoincidence := helpers.CheckCoincidenceProduction(SolicitudProduccionAlert, idTipoProduccion, idTercero); errCoincidence == nil {
 				idStr := fmt.Sprintf("%v", SolicitudProduccionPut["Id"])
 				fmt.Println(idStr)
-				if resultadoPutSolicitudDocente, errPut := models.PutSolicitudDocente(SolicitudProduccionPut, idStr); errPut == nil {
+				if resultadoPutSolicitudDocente, errPut := helpers.PutSolicitudDocente(SolicitudProduccionPut, idStr); errPut == nil {
 					resultado = resultadoPutSolicitudDocente
 					c.Data["json"] = resultado
 				} else {
@@ -85,7 +85,7 @@ func (c *SolicitudProduccionController) PutResultadoSolicitud() {
 	fmt.Println("Id es: " + idStr)
 	var SolicitudProduccion map[string]interface{}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &SolicitudProduccion); err == nil {
-		if SolicitudProduccionResult, errPuntaje := models.GenerateResult(SolicitudProduccion); errPuntaje == nil {
+		if SolicitudProduccionResult, errPuntaje := helpers.GenerateResult(SolicitudProduccion); errPuntaje == nil {
 			c.Data["json"] = SolicitudProduccionResult
 		} else {
 			logs.Error(SolicitudProduccionResult)
@@ -120,7 +120,7 @@ func (c *SolicitudProduccionController) PostSolicitudEvaluacionCoincidencia() {
 	fmt.Println("Id Solicitud Coincidencia: ", idSolicitudCoincidencia)
 
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &SolicitudProduccion); err == nil {
-		if SolicitudProduccionClone, errClone := models.GenerateEvaluationsCloning(SolicitudProduccion, idSolicitud, idSolicitudCoincidencia, idTercero); errClone == nil {
+		if SolicitudProduccionClone, errClone := helpers.GenerateEvaluationsCloning(SolicitudProduccion, idSolicitud, idSolicitudCoincidencia, idTercero); errClone == nil {
 			if len(SolicitudProduccionClone) > 0 {
 				resultado = SolicitudProduccion
 				c.Data["json"] = resultado
