+++ helpers/inscripciones.go
@@ -1,12 +1,84 @@
-package models
+package helpers
 
 import (
 	"fmt"
 
 	"github.com/astaxie/beego"
+	"github.com/udistrital/sga_mid_inscripcion/utils"
 	"github.com/udistrital/utils_oas/request"
 )
 
+func SetInactivo(url string) (exito bool) {
+	exito = false
+	var payload1 map[string]interface{}
+	fmt.Println(url)
+	errGet := request.GetJson(url, &payload1)
+	if errGet == nil {
+		fmt.Println(payload1)
+		var idDisable string = ""
+		var body map[string]interface{}
+		if payload1["Id"] != nil {
+			fmt.Println("is by id only")
+			idDisable = fmt.Sprintf("%v", payload1["Id"])
+			body = payload1
+		}
+		if payload1["Data"] != nil {
+			fmt.Println("is is inside data")
+			idDisable = fmt.Sprintf("%v", payload1["Data"].(map[string]interface{})["Id"])
+			body = payload1["Data"].(map[string]interface{})
+		}
+
+		fmt.Println("id is:", idDisable)
+
+		if idDisable != "" {
+			body["Activo"] = false
+			fmt.Println("body is:", body)
+			var payload2 map[string]interface{}
+			errSet := request.SendJson(url, "PUT", &payload2, body)
+			if errSet == nil {
+				if payload2["Id"] != nil {
+					if fmt.Sprintf("%v", payload2["Id"]) == idDisable {
+						exito = true
+					} else {
+						exito = false
+					}
+				} else if payload1["Data"] != nil {
+					if fmt.Sprintf("%v", payload2["Data"].(map[string]interface{})["Id"]) == idDisable {
+						exito = true
+					} else {
+						exito = false
+					}
+				} else {
+					exito = false
+				}
+			} else {
+				exito = false
+			}
+		} else {
+			exito = false
+		}
+	} else {
+		exito = false
+	}
+
+	return exito
+}
+
+// IdInfoCompTercero is ...
+func IdInfoCompTercero(grupo string, codAbrev string) (Id string, ok bool) {
+	var resp []map[string]interface{}
+	errResp := request.GetJson("http://"+beego.AppConfig.String("TercerosService")+"info_complementaria?query=GrupoInfoComplementariaId__Id:"+grupo+",CodigoAbreviacion:"+codAbrev+"&fields=Id", &resp)
+	if errResp == nil && fmt.Sprintf("%v", resp) != "[map[]]" {
+		Id = fmt.Sprintf("%v", resp[0]["Id"].(float64))
+		ok = true
+	} else {
+		Id = "0"
+		ok = false
+	}
+	return Id, ok
+}
+
+//Verificar estado de lso recibos ...
 func VerificarRecibos(personaId string, periodoId string) ( resultadoAuxResponse map[string]interface{}, Error string) {
 	var Inscripciones []map[string]interface{}
 	var ReciboXML map[string]interface{}
@@ -40,7 +112,7 @@ func VerificarRecibos(personaId string, periodoId string) ( resultadoAuxResponse
 									Estado = "Pago"
 								} else {
 									//Verifica si el recibo está vencido o no
-									ATiempo, err := VerificarFechaLimite(FechaLimite)
+									ATiempo, err := utils.VerificarFechaLimite(FechaLimite)
 									if err == nil {
 										if ATiempo {
 											Estado = "Pendiente pago"
