+++ helpers/produccion_academica.go
@@ -1,7 +1,8 @@
-package models
+package helpers
 
 import (
 	"fmt"
+
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
 	"github.com/udistrital/utils_oas/request"
