+++ utils/SendJsonHeader.go
@@ -1,4 +1,4 @@
-package models
+package utils
 
 import (
 	"bytes"
